+++ seed/src/lib.rs
@@ -19,24 +19,20 @@
 #[macro_use]
 extern crate async_trait;
 
-use std::{
-    collections::{HashMap, HashSet},
-    io,
-    net::SocketAddr,
-    path::PathBuf,
-    time::Duration,
-    vec,
-};
+pub mod event;
+pub mod handle;
+pub mod project;
+pub mod signer;
+
+use std::{collections::HashSet, net::SocketAddr, path::PathBuf, time::Duration, vec};
 
 use futures::{channel::mpsc as chan, select, sink::SinkExt as _, stream::StreamExt as _};
 use thiserror::Error;
 
-use radicle_keystore::sign::ed25519;
-
 use librad::{
     git,
     keys,
-    meta::{entity, project, project::ProjectInfo, user::User},
+    meta::{self, entity, user::User},
     net::{
         discovery,
         gossip,
@@ -47,6 +43,13 @@ use librad::{
     paths,
     peer::PeerId,
     uri::RadUrn,
+};
+
+pub use crate::{
+    event::Event,
+    handle::{NodeError, NodeHandle, Request},
+    project::Project,
+    signer::Signer,
 };
 
 #[derive(Error, Debug)]
@@ -76,45 +79,6 @@ pub enum Error {
     Channel(#[from] chan::SendError),
 }
 
-#[derive(Clone)]
-pub struct Signer {
-    key: keys::SecretKey,
-}
-
-impl Signer {
-    pub fn new<R: io::Read>(mut r: R) -> Result<Self, io::Error> {
-        use radicle_keystore::SecretKeyExt;
-
-        let mut bytes = Vec::new();
-        r.read_to_end(&mut bytes)?;
-
-        let sbytes: keys::SecStr = bytes.into();
-        match keys::SecretKey::from_bytes_and_meta(sbytes, &()) {
-            Ok(key) => Ok(Self { key }),
-            Err(err) => Err(io::Error::new(io::ErrorKind::InvalidData, err)),
-        }
-    }
-}
-
-#[async_trait]
-impl ed25519::Signer for Signer {
-    type Error = std::convert::Infallible;
-
-    fn public_key(&self) -> ed25519::PublicKey {
-        self.key.public_key()
-    }
-
-    async fn sign(&self, data: &[u8]) -> Result<ed25519::Signature, Self::Error> {
-        <keys::SecretKey as ed25519::Signer>::sign(&self.key, data).await
-    }
-}
-
-impl keys::AsPKCS8 for Signer {
-    fn as_pkcs8(&self) -> Vec<u8> {
-        self.key.as_pkcs8()
-    }
-}
-
 /// Seed operational mode.
 #[derive(Debug)]
 pub enum Mode {
@@ -134,67 +98,6 @@ impl Mode {
             Mode::TrackUrns(ref urns) => urns.contains(urn),
             Mode::TrackPeers(ref peers) => peers.contains(peer),
         }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct Project {
-    pub urn: RadUrn,
-    pub name: String,
-    pub description: Option<String>,
-    pub maintainers: HashSet<RadUrn>,
-}
-
-/// An event generated by the seed node.
-#[derive(Debug, Clone)]
-pub enum Event {
-    /// The seed node is listening for peer connections.
-    Listening(SocketAddr),
-    /// A peer has connected.
-    PeerConnected {
-        peer_id: PeerId,
-        urn: Option<RadUrn>,
-        name: Option<String>,
-    },
-    /// A peer has disconnected.
-    PeerDisconnected(PeerId),
-    /// A project has been tracked from a peer.
-    ProjectTracked(Project, PeerId),
-}
-
-impl Event {
-    async fn peer_connected(peer_id: PeerId, api: &PeerApi<Signer>) -> Result<Self, Error> {
-        let user = self::guess_user(peer_id, api).await?;
-        let user = user.as_ref();
-
-        Ok(Self::PeerConnected {
-            peer_id,
-            urn: user.map(|u| u.urn()),
-            name: user.map(|u| u.name().to_owned()),
-        })
-    }
-
-    async fn project_tracked(
-        urn: RadUrn,
-        provider: PeerId,
-        api: &PeerApi<Signer>,
-    ) -> Result<Self, Error> {
-        let proj = api
-            .with_storage({
-                let urn = urn.clone();
-                move |s| s.metadata_of::<ProjectInfo, _>(&urn, provider)
-            })
-            .await??;
-
-        Ok(Event::ProjectTracked(
-            Project {
-                urn: urn.clone(),
-                maintainers: proj.maintainers().clone(),
-                name: proj.name().to_owned(),
-                description: proj.description().to_owned(),
-            },
-            provider,
-        ))
     }
 }
 
@@ -223,50 +126,6 @@ impl Default for NodeConfig {
     }
 }
 
-/// An error returned by the [`NodeHandle`].
-#[derive(Debug, Error)]
-pub enum NodeError {
-    #[error("request failed: the node disconnected")]
-    RequestFailed,
-}
-
-/// Handle used to interact with the seed node.
-pub struct NodeHandle {
-    channel: chan::UnboundedSender<Request>,
-}
-
-impl NodeHandle {
-    /// Get all local projects.
-    pub async fn get_projects(&mut self) -> Result<Vec<Project>, NodeError> {
-        let (tx, mut rx) = chan::channel(1);
-        self.channel
-            .send(Request::GetProjects(tx))
-            .await
-            .map_err(|_| NodeError::RequestFailed)?;
-
-        rx.next().await.ok_or(NodeError::RequestFailed)
-    }
-
-    /// Get currently connected peers.
-    pub async fn get_peers(&mut self) -> Result<HashMap<PeerId, SocketAddr>, NodeError> {
-        let (tx, mut rx) = chan::channel(1);
-        self.channel
-            .send(Request::GetPeers(tx))
-            .await
-            .map_err(|_| NodeError::RequestFailed)?;
-
-        rx.next().await.ok_or(NodeError::RequestFailed)
-    }
-}
-
-/// User request to the seed node.
-enum Request {
-    /// Get local projects.
-    GetProjects(chan::Sender<Vec<Project>>),
-    /// Get connected peers.
-    GetPeers(chan::Sender<HashMap<PeerId, SocketAddr>>),
-}
-
 /// Seed node instance.
 pub struct Node {
     config: NodeConfig,
@@ -290,9 +149,7 @@ impl Node {
 
     /// Create a new handle.
     pub fn handle(&self) -> NodeHandle {
-        NodeHandle {
-            channel: self.handle.clone(),
-        }
+        NodeHandle::new(self.handle.clone())
     }
 
     /// Run the seed node. This function runs indefinitely until a fatal error
@@ -358,7 +215,7 @@ impl Node {
                 reply.send(peers).await?;
             },
             Request::GetProjects(mut reply) => {
-                let projs = self::get_projects(api).await?;
+                let projs = project::get_projects(api).await?;
                 reply.send(projs).await?;
             },
         }
@@ -424,7 +281,7 @@ impl Node {
             let urn = urn.clone();
             api.with_storage(move |storage| {
                 storage
-                    .clone_repo::<project::ProjectInfo, _>(url, addr_hints)
+                    .clone_repo::<meta::project::ProjectInfo, _>(url, addr_hints)
                     .and_then(|_| storage.track(&urn, &peer_id))
             })
         }
@@ -507,32 +364,6 @@ async fn guess_user(
             }
         }
         Ok(None)
-    })
-    .await?
-}
-
-/// Get all local projects.
-async fn get_projects(api: &PeerApi<Signer>) -> Result<Vec<Project>, Error> {
-    api.with_storage(|s| {
-        let projs = s
-            .all_metadata()?
-            .flat_map(|meta| {
-                let meta = meta.ok()?;
-
-                meta.try_map(|info| match info {
-                    entity::data::EntityInfo::Project(info) => Some(info),
-                    _ => None,
-                })
-                .map(|meta| Project {
-                    urn: meta.urn(),
-                    name: meta.name().to_owned(),
-                    description: meta.description().to_owned(),
-                    maintainers: meta.maintainers().clone(),
-                })
-            })
-            .collect::<Vec<Project>>();
-
-        Ok(projs)
     })
     .await?
 }
